requirements.txt
absl-py==0.15.0
actionlib==1.12.1
alabaster==0.7.8
angles==1.9.12
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
astroid==1.6.0
astunparse==1.6.3
attrs==17.4.0
Babel==2.4.0
bcrypt==3.2.0
beautifulsoup4==4.6.0
bleach==2.1.2
bondpy==1.8.5
Brlapi==0.6.6
cached-property==1.5.2
cachetools==4.2.4
camera_calibration==1.15.2
camera_calibration_parsers==1.11.13
catkin==0.7.29
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloudpickle==2.2.0
cmake==3.25.0
colorama==0.3.7
command-not-found==0.3
controller_manager==0.18.4
controller_manager_msgs==0.18.4
coremltools==6.1
cryptography==3.2.1
cupshelpers==1.0
cv_bridge==1.13.1
cycler==0.10.0
dataclasses==0.8
decorator==4.1.2
defer==1.0.6
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
dm-tree==0.1.8
docker==4.3.1
docker-compose==1.27.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
dynamic_reconfigure==1.6.5
dynamixel_sdk==3.7.51
entrypoints==0.2.3.post1
f.lux indicator applet==1.1.11rc0
flake8==3.5.0
flatbuffers==1.12
gast==0.4.0
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.6
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.17
gennodejs==2.0.1
genpy==0.6.16
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
html5lib==0.999999999
httplib2==0.9.2
idna==2.6
image_geometry==1.13.1
imagesize==0.7.1
importlib-metadata==4.8.3
imutils==0.5.4
interactive-markers==1.11.5
iotop==0.6
ipykernel==4.8.2
ipython==5.5.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
isort==4.3.4
jedi==0.11.1
Jinja2==2.10
joblib==0.11
joint_state_publisher==1.12.15
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-console==5.2.0
jupyter-core==4.4.0
kazam==1.4.5
kdl-parser-py==1.13.3
Keras==2.3.1
Keras-Applications==1.0.8
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
laser_geometry==1.6.7
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.3.1
libclang==11.1.0
logilab-common==1.4.1
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.7
MarkupSafe==1.0
matplotlib==3.3.4
mccabe==0.6.1
message_filters==1.14.13
mistune==0.8.3
monai==0.8.1
moveit-core==1.0.11
moveit_ros_planning_interface==1.0.11
moveit_ros_visualization==1.0.11
mpmath==1.2.1
natsort==8.2.0
nbconvert==5.3.1
nbformat==4.4.0
netifaces==0.10.4
netron==6.4.0
nose==1.3.7
nose-parameterized==0.3.4
notebook==5.2.2
numexpr==2.6.4
numpy==1.19.5
numpydoc==0.7.0
nvidia-ml-py==7.352.0
oauth==1.0.1
oauthlib==3.1.1
olefile==0.45.1
onnx==1.4.1
opt-einsum==3.3.0
packaging==21.3
pandas==0.22.0
pandocfilters==1.4.2
paramiko==2.7.2
parso==0.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==8.4.0
pluggy==0.6.0
prompt-toolkit==1.0.15
protobuf==3.19.1
psutil==5.4.2
py==1.5.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.3.1
pycrypto==2.6.1
pycups==1.9.73
pydot==1.2.3
pyflakes==1.6.0
Pygments==2.2.0
PyGObject==3.26.1
pygpu==0.7.6
pylint==1.8.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyRFC3339==1.0
pytest==3.3.2
python-apt==1.6.6
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.15.0
python_qt_binding==0.4.4
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
reportlab==3.4.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
resource_retriever==1.12.7
roman==2.0.0
rope==0.10.5
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.14
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.3
rqt_gui_py==0.5.3
rqt_image_view==0.4.16
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.3
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rsa==4.7.2
rviz==1.13.29
scikit-learn==0.19.1
scipy==1.4.1
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
sensor-msgs==1.12.8
simplegeneric==0.8.1
simplejson==3.13.2
six==1.15.0
smach==2.0.1
smach_ros==2.0.1
smclib==1.8.5
Sphinx==1.6.7
spyder==3.2.6
srdfdom==0.5.2
ssh-import-id==5.7
sympy==1.9
system-service==0.3
systemd-python==234
tables==3.4.2
tb-nightly==2.6.0a20210806
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-addons==0.10.0
tensorflow-docs @ git+https://github.com/tensorflow/docs@ed4a55140445dfd33ed8ac5c9b21b9fa86ec9b44
tensorflow-estimator==2.5.0
tensorflow-probability==0.19.0
termcolor==1.1.0
terminado==0.7
testpath==0.3.1
texttable==1.6.3
tf==1.12.1
tf-estimator-nightly==2.7.0.dev2021092408
tf-nightly-gpu==2.7.0.dev20210806
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
Theano==1.0.4
topic_tools==1.14.13
torch==1.10.2
torchcam==0.3.2
tornado==4.5.3
tqdm==4.62.3
traitlets==4.3.2
typeguard==2.13.0
typing==3.7.4.3
typing-extensions==3.7.4.3
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urdfdom-py==0.4.6
urllib3==1.22
usb-creator==0.3.3
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.1.7
webencodings==0.5
websocket-client==0.57.0
Werkzeug==2.0.3
wrapt==1.12.1
xacro==1.13.18
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
